backend/requirements.txt
google-cloud-resource-manager
google-cloud-asset
flask